Problem: app utils triggers import cycle easily
mv to testutil and config
