+++ app/app_test.go
@@ -1,9 +1,11 @@
-package app
+package app_test
 
 import (
 	"os"
 	"testing"
 
+	"github.com/evmos/ethermint/app"
+	"github.com/evmos/ethermint/testutil"
 	"github.com/stretchr/testify/require"
 
 	"cosmossdk.io/log"
@@ -15,18 +17,18 @@ import (
 
 func TestEthermintAppExport(t *testing.T) {
 	db := dbm.NewMemDB()
-	app := SetupWithDB(false, nil, db)
-	app.Commit()
+	ethApp := testutil.SetupWithDB(false, nil, db)
+	ethApp.Commit()
 
 	// Making a new app object with the db, so that initchain hasn't been called
-	app2 := NewEthermintApp(
+	ethApp2 := app.NewEthermintApp(
 		log.NewLogger(os.Stdout),
 		db,
 		nil,
 		true,
-		simtestutil.NewAppOptionsWithFlagHome(DefaultNodeHome),
-		baseapp.SetChainID(ChainID),
+		simtestutil.NewAppOptionsWithFlagHome(app.DefaultNodeHome),
+		baseapp.SetChainID(testutil.ChainID),
 	)
-	_, err := app2.ExportAppStateAndValidators(false, []string{}, []string{})
+	_, err := ethApp2.ExportAppStateAndValidators(false, []string{}, []string{})
 	require.NoError(t, err, "ExportAppStateAndValidators should not have an error")
 }
