--- app/utils.go
@@ -1,375 +1,0 @@
-// Copyright 2021 Evmos Foundation
-// This file is part of Evmos' Ethermint library.
-//
-// The Ethermint library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// The Ethermint library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with the Ethermint library. If not, see https://github.com/evmos/ethermint/blob/main/LICENSE
-package app
-
-import (
-	"encoding/json"
-	"math/rand"
-	"time"
-
-	"cosmossdk.io/log"
-	sdkmath "cosmossdk.io/math"
-	"cosmossdk.io/x/evidence"
-	feegrantmodule "cosmossdk.io/x/feegrant/module"
-	"cosmossdk.io/x/upgrade"
-	abci "github.com/cometbft/cometbft/abci/types"
-	cmtproto "github.com/cometbft/cometbft/proto/tendermint/types"
-	tmtypes "github.com/cometbft/cometbft/types"
-	dbm "github.com/cosmos/cosmos-db"
-	"github.com/cosmos/cosmos-sdk/baseapp"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/codec"
-	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
-	cryptocodec "github.com/cosmos/cosmos-sdk/crypto/codec"
-	"github.com/cosmos/cosmos-sdk/crypto/keys/ed25519"
-	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
-	"github.com/cosmos/cosmos-sdk/server"
-	"github.com/cosmos/cosmos-sdk/testutil/mock"
-	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/module"
-	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-	"github.com/cosmos/cosmos-sdk/x/auth/vesting"
-	authzmodule "github.com/cosmos/cosmos-sdk/x/authz/module"
-	"github.com/cosmos/cosmos-sdk/x/bank"
-	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	"github.com/cosmos/cosmos-sdk/x/consensus"
-	"github.com/cosmos/cosmos-sdk/x/crisis"
-	"github.com/cosmos/cosmos-sdk/x/gov"
-	govclient "github.com/cosmos/cosmos-sdk/x/gov/client"
-	"github.com/cosmos/cosmos-sdk/x/mint"
-	"github.com/cosmos/cosmos-sdk/x/params"
-	paramsclient "github.com/cosmos/cosmos-sdk/x/params/client"
-	"github.com/cosmos/cosmos-sdk/x/slashing"
-	"github.com/cosmos/cosmos-sdk/x/staking"
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	"github.com/cosmos/ibc-go/modules/capability"
-	"github.com/cosmos/ibc-go/v8/modules/apps/transfer"
-	ibc "github.com/cosmos/ibc-go/v8/modules/core"
-	ibctm "github.com/cosmos/ibc-go/v8/modules/light-clients/07-tendermint"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/evmos/ethermint/crypto/ethsecp256k1"
-	"github.com/evmos/ethermint/encoding"
-	ethermint "github.com/evmos/ethermint/types"
-	"github.com/evmos/ethermint/x/evm"
-	evmtypes "github.com/evmos/ethermint/x/evm/types"
-	"github.com/evmos/ethermint/x/feemarket"
-
-	"github.com/cosmos/cosmos-sdk/x/auth"
-	distr "github.com/cosmos/cosmos-sdk/x/distribution"
-	"github.com/cosmos/cosmos-sdk/x/genutil"
-	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
-)
-
-type GenesisState map[string]json.RawMessage
-
-// DefaultConsensusParams defines the default Tendermint consensus params used in
-// EthermintApp testing.
-var DefaultConsensusParams = &cmtproto.ConsensusParams{
-	Block: &cmtproto.BlockParams{
-		MaxBytes: 1048576,
-		MaxGas:   81500000, // default limit
-	},
-	Evidence: &cmtproto.EvidenceParams{
-		MaxAgeNumBlocks: 302400,
-		MaxAgeDuration:  504 * time.Hour, // 3 weeks is the max duration
-		MaxBytes:        10000,
-	},
-	Validator: &cmtproto.ValidatorParams{
-		PubKeyTypes: []string{
-			tmtypes.ABCIPubKeyTypeEd25519,
-		},
-	},
-}
-
-// Setup initializes a new EthermintApp. A Nop logger is set in EthermintApp.
-func Setup(isCheckTx bool, patch func(*EthermintApp, GenesisState) GenesisState) *EthermintApp {
-	return SetupWithDB(isCheckTx, patch, dbm.NewMemDB())
-}
-
-func SetupWithOpts(
-	isCheckTx bool,
-	patch func(*EthermintApp, GenesisState) GenesisState,
-	appOptions simtestutil.AppOptionsMap,
-) *EthermintApp {
-	return SetupWithDBAndOpts(isCheckTx, patch, dbm.NewMemDB(), appOptions)
-}
-
-const ChainID = "ethermint_9000-1"
-
-func SetupWithDB(isCheckTx bool, patch func(*EthermintApp, GenesisState) GenesisState, db dbm.DB) *EthermintApp {
-	return SetupWithDBAndOpts(isCheckTx, patch, db, nil)
-}
-
-// SetupWithDBAndOpts initializes a new EthermintApp. A Nop logger is set in EthermintApp.
-func SetupWithDBAndOpts(
-	isCheckTx bool,
-	patch func(*EthermintApp, GenesisState) GenesisState,
-	db dbm.DB,
-	appOptions simtestutil.AppOptionsMap,
-) *EthermintApp {
-	if appOptions == nil {
-		appOptions = make(simtestutil.AppOptionsMap, 0)
-	}
-	appOptions[server.FlagInvCheckPeriod] = 5
-	appOptions[flags.FlagHome] = DefaultNodeHome
-	app := NewEthermintApp(log.NewNopLogger(),
-		db,
-		nil,
-		true,
-		appOptions,
-		baseapp.SetChainID(ChainID),
-	)
-
-	if !isCheckTx {
-		// init chain must be called to stop deliverState from being nil
-		genesisState := NewTestGenesisState(app.AppCodec(), app.DefaultGenesis())
-		if patch != nil {
-			genesisState = patch(app, genesisState)
-		}
-
-		stateBytes, err := json.MarshalIndent(genesisState, "", " ")
-		if err != nil {
-			panic(err)
-		}
-
-		// Initialize the chain
-		consensusParams := DefaultConsensusParams
-		initialHeight := app.LastBlockHeight() + 1
-		consensusParams.Abci = &cmtproto.ABCIParams{VoteExtensionsEnableHeight: initialHeight}
-		if _, err := app.InitChain(
-			&abci.RequestInitChain{
-				ChainId:         ChainID,
-				Validators:      []abci.ValidatorUpdate{},
-				ConsensusParams: consensusParams,
-				AppStateBytes:   stateBytes,
-				InitialHeight:   initialHeight,
-			},
-		); err != nil {
-			panic(err)
-		}
-	}
-	if _, err := app.FinalizeBlock(&abci.RequestFinalizeBlock{
-		Height: app.LastBlockHeight() + 1,
-		Hash:   app.LastCommitID().Hash,
-	}); err != nil {
-		panic(err)
-	}
-	return app
-}
-
-// RandomGenesisAccounts is used by the auth module to create random genesis accounts in simulation when a genesis.json is not specified.
-// In contrast, the default auth module's RandomGenesisAccounts implementation creates only base accounts and vestings accounts.
-func RandomGenesisAccounts(simState *module.SimulationState) authtypes.GenesisAccounts {
-	emptyCodeHash := crypto.Keccak256(nil)
-	genesisAccs := make(authtypes.GenesisAccounts, len(simState.Accounts))
-	for i, acc := range simState.Accounts {
-		bacc := authtypes.NewBaseAccountWithAddress(acc.Address)
-
-		ethacc := &ethermint.EthAccount{
-			BaseAccount: bacc,
-			CodeHash:    common.BytesToHash(emptyCodeHash).String(),
-		}
-		genesisAccs[i] = ethacc
-	}
-
-	return genesisAccs
-}
-
-// RandomAccounts creates random accounts with an ethsecp256k1 private key
-// TODO: replace secp256k1.GenPrivKeyFromSecret() with similar function in go-ethereum
-func RandomAccounts(r *rand.Rand, n int) []simtypes.Account {
-	accs := make([]simtypes.Account, n)
-
-	for i := 0; i < n; i++ {
-		// don't need that much entropy for simulation
-		privkeySeed := make([]byte, 15)
-		_, _ = r.Read(privkeySeed)
-
-		prv := secp256k1.GenPrivKeyFromSecret(privkeySeed)
-		ethPrv := &ethsecp256k1.PrivKey{}
-		_ = ethPrv.UnmarshalAmino(prv.Bytes()) // UnmarshalAmino simply copies the bytes and assigns them to ethPrv.Key
-		accs[i].PrivKey = ethPrv
-		accs[i].PubKey = accs[i].PrivKey.PubKey()
-		accs[i].Address = sdk.AccAddress(accs[i].PubKey.Address())
-
-		accs[i].ConsKey = ed25519.GenPrivKeyFromSecret(privkeySeed)
-	}
-
-	return accs
-}
-
-// StateFn returns the initial application state using a genesis or the simulation parameters.
-// It is a wrapper of simapp.AppStateFn to replace evm param EvmDenom with staking param BondDenom.
-func StateFn(app *EthermintApp) simtypes.AppStateFn {
-	var bondDenom string
-	return simtestutil.AppStateFnWithExtendedCbs(
-		app.AppCodec(),
-		app.SimulationManager(),
-		app.DefaultGenesis(),
-		func(moduleName string, genesisState interface{}) {
-			if moduleName == stakingtypes.ModuleName {
-				stakingState := genesisState.(*stakingtypes.GenesisState)
-				bondDenom = stakingState.Params.BondDenom
-			}
-		},
-		func(rawState map[string]json.RawMessage) {
-			evmStateBz, ok := rawState[evmtypes.ModuleName]
-			if !ok {
-				panic("evm genesis state is missing")
-			}
-
-			evmState := new(evmtypes.GenesisState)
-			app.AppCodec().MustUnmarshalJSON(evmStateBz, evmState)
-
-			// we should replace the EvmDenom with BondDenom
-			evmState.Params.EvmDenom = bondDenom
-
-			// change appState back
-			rawState[evmtypes.ModuleName] = app.AppCodec().MustMarshalJSON(evmState)
-		},
-	)
-}
-
-// NewTestGenesisState generate genesis state with single validator
-func NewTestGenesisState(codec codec.Codec, genesisState map[string]json.RawMessage) GenesisState {
-	privVal := mock.NewPV()
-	pubKey, err := privVal.GetPubKey()
-	if err != nil {
-		panic(err)
-	}
-	// create validator set with single validator
-	validator := tmtypes.NewValidator(pubKey, 1)
-	valSet := tmtypes.NewValidatorSet([]*tmtypes.Validator{validator})
-
-	// generate genesis account
-	senderPrivKey := secp256k1.GenPrivKey()
-	acc := authtypes.NewBaseAccount(senderPrivKey.PubKey().Address().Bytes(), senderPrivKey.PubKey(), 0, 0)
-	balance := banktypes.Balance{
-		Address: acc.GetAddress().String(),
-		Coins:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdkmath.NewInt(100000000000000))),
-	}
-
-	return genesisStateWithValSet(codec, genesisState, valSet, []authtypes.GenesisAccount{acc}, balance)
-}
-
-func genesisStateWithValSet(codec codec.Codec, genesisState GenesisState,
-	valSet *tmtypes.ValidatorSet, genAccs []authtypes.GenesisAccount,
-	balances ...banktypes.Balance,
-) GenesisState {
-	// set genesis accounts
-	authGenesis := authtypes.NewGenesisState(authtypes.DefaultParams(), genAccs)
-	genesisState[authtypes.ModuleName] = codec.MustMarshalJSON(authGenesis)
-
-	validators := make([]stakingtypes.Validator, 0, len(valSet.Validators))
-	delegations := make([]stakingtypes.Delegation, 0, len(valSet.Validators))
-
-	bondAmt := sdk.DefaultPowerReduction
-
-	for _, val := range valSet.Validators {
-		pk, err := cryptocodec.FromCmtPubKeyInterface(val.PubKey)
-		if err != nil {
-			panic(err)
-		}
-		pkAny, err := codectypes.NewAnyWithValue(pk)
-		if err != nil {
-			panic(err)
-		}
-		validator := stakingtypes.Validator{
-			OperatorAddress:   sdk.ValAddress(val.Address).String(),
-			ConsensusPubkey:   pkAny,
-			Jailed:            false,
-			Status:            stakingtypes.Bonded,
-			Tokens:            bondAmt,
-			DelegatorShares:   sdkmath.LegacyOneDec(),
-			Description:       stakingtypes.Description{},
-			UnbondingHeight:   int64(0),
-			UnbondingTime:     time.Unix(0, 0).UTC(),
-			Commission:        stakingtypes.NewCommission(sdkmath.LegacyOneDec(), sdkmath.LegacyOneDec(), sdkmath.LegacyOneDec()),
-			MinSelfDelegation: sdkmath.ZeroInt(),
-		}
-		validators = append(validators, validator)
-		delegation := stakingtypes.NewDelegation(genAccs[0].GetAddress().String(), sdk.ValAddress(val.Address).String(), sdkmath.LegacyOneDec())
-		delegations = append(delegations, delegation)
-	}
-	// set validators and delegations
-	stakingGenesis := stakingtypes.NewGenesisState(stakingtypes.DefaultParams(), validators, delegations)
-	genesisState[stakingtypes.ModuleName] = codec.MustMarshalJSON(stakingGenesis)
-
-	totalSupply := sdk.NewCoins()
-	for _, b := range balances {
-		// add genesis acc tokens to total supply
-		totalSupply = totalSupply.Add(b.Coins...)
-	}
-
-	for range delegations {
-		// add delegated tokens to total supply
-		totalSupply = totalSupply.Add(sdk.NewCoin(sdk.DefaultBondDenom, bondAmt))
-	}
-
-	// add bonded amount to bonded pool module account
-	balances = append(balances, banktypes.Balance{
-		Address: authtypes.NewModuleAddress(stakingtypes.BondedPoolName).String(),
-		Coins:   sdk.Coins{sdk.NewCoin(sdk.DefaultBondDenom, bondAmt)},
-	})
-
-	// update total supply
-	bankGenesis := banktypes.NewGenesisState(
-		banktypes.DefaultGenesisState().Params,
-		balances,
-		totalSupply,
-		[]banktypes.Metadata{},
-		[]banktypes.SendEnabled{},
-	)
-	genesisState[banktypes.ModuleName] = codec.MustMarshalJSON(bankGenesis)
-
-	return genesisState
-}
-
-var ModuleBasicsForTest = module.NewBasicManager(
-	auth.AppModuleBasic{},
-	genutil.NewAppModuleBasic(genutiltypes.DefaultMessageValidator),
-	bank.AppModuleBasic{},
-	capability.AppModuleBasic{},
-	staking.AppModuleBasic{},
-	mint.AppModuleBasic{},
-	distr.AppModuleBasic{},
-	gov.NewAppModuleBasic([]govclient.ProposalHandler{paramsclient.ProposalHandler}),
-	params.AppModuleBasic{},
-	crisis.AppModuleBasic{},
-	slashing.AppModuleBasic{},
-	ibc.AppModuleBasic{},
-	ibctm.AppModuleBasic{},
-	authzmodule.AppModuleBasic{},
-	feegrantmodule.AppModuleBasic{},
-	upgrade.AppModuleBasic{},
-	evidence.AppModuleBasic{},
-	transfer.AppModuleBasic{},
-	vesting.AppModuleBasic{},
-	consensus.AppModuleBasic{},
-	// Ethermint modules
-	evm.AppModuleBasic{},
-	feemarket.AppModuleBasic{},
-)
-
-func MakeConfigForTest() ethermint.EncodingConfig {
-	config := encoding.MakeConfig()
-	ModuleBasicsForTest.RegisterLegacyAminoCodec(config.Amino)
-	ModuleBasicsForTest.RegisterInterfaces(config.InterfaceRegistry)
-	return config
-}
